Cleanup placement of new traits
